+++ src/Products/urban/base.py
@@ -709,14 +709,14 @@ class UrbanBase(object):
     def listSolicitOpinionsTo(self, unless=[]):
         return removeItems(list(self.getValuesForTemplate('solicitOpinionsTo')), unless)
 
-    security.declarePublic('getFirstAdministrativeSfolderManager')
-    def getFirstAdministrativeSfolderManager(self, administrativeVocabularyId='agent-administratif'):
-        FASFM = None
+    security.declarePublic('getFirstGradeIdSfolderManager')
+    def getFirstGradeIdSfolderManager(self, gradeId=''):
+        folderManager = None
         found = False
         i = 0
         while not found and i < len(self.getFoldermanagers()):
-            if self.getFoldermanagers()[i].getGrade() == administrativeVocabularyId:
+            if self.getFoldermanagers()[i].getGrade() == gradeId:
                 found = True
-                FASFM = self.getFoldermanagers()[i]
+                folderManager = self.getFoldermanagers()[i]
             i = i + 1
-        return FASFM
+        return folderManager